feat: contribution with kindness to, no code no life. I have made several improvements to the project. All changes are documented in the doc folder
