+++ Desarrollo web/WebPersonal/src/constants/languages.js
@@ -1,0 +1,12 @@
+const languages = [
+    { name: 'Python', icon: 'https://upload.wikimedia.org/wikipedia/commons/c/c3/Python-logo-notext.svg' },
+    { name: 'Java', icon: 'https://img.icons8.com/?size=100&id=13679&format=png&color=000000' },
+    { name: 'TypeScript', icon: 'https://img.icons8.com/?size=100&id=wpZmKzk11AzJ&format=png&color=000000' },
+    { name: 'JavaScript', icon: 'https://img.icons8.com/?size=100&id=108784&format=png&color=000000' },
+    { name: 'HTML', icon: 'https://img.icons8.com/?size=100&id=20909&format=png&color=000000' },
+    { name: 'CSS', icon: 'https://img.icons8.com/?size=100&id=21278&format=png&color=000000' },
+    { name: 'Arduino', icon: 'https://img.icons8.com/?size=100&id=13444&format=png&color=000000' },
+    { name: 'C', icon: 'https://img.icons8.com/?size=512&id=40670&format=png' },
+];
+
+export { languages };
