+++ Desarrollo web/WebPersonal/src/functionality/particles/app.js
@@ -5,7 +5,7 @@
 /* To load a config file (particles.json) you need to host this demo (MAMP/WAMP/local)... */
 
 /* particlesJS.load(@dom-id, @path-json, @callback (optional)); */
-particlesJS.load('particles-js', 'particles.json', function() {
+particlesJS.load('particles-js', '../../src/assets/json/particles.json', function () {
   console.log('callback - particles.js config loaded');
 });
 
@@ -13,7 +13,7 @@ particlesJS.load('particles-js', 'particles.json', function() {
 /* Otherwise just put the config content (json): */
 
 particlesJS('particles-js',
-  
+
   {
     "particles": {
       "number": {
