+++ Desarrollo web/WebPersonal/src/style-components/animations/typing.css
@@ -1,0 +1,13 @@
+@keyframes blink {
+    0% {
+        opacity: 1;
+    }
+
+    50% {
+        opacity: 0;
+    }
+
+    100% {
+        opacity: 1;
+    }
+}
