+++ Desarrollo web/WebPersonal/src/style-components/main/stack.css
@@ -1,0 +1,29 @@
+.stack__div {
+    display: flex;
+    flex-direction: column;
+    gap: calc(var(--gap-uniform) * 6);
+
+    h3 {
+        padding-bottom: calc(var(--padding-uniform) * 3);
+    }
+}
+
+.stack__div--skills {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+    gap: var(--gap-uniform);
+
+    width: 100%;
+    margin: auto;
+
+    .stack-skill {
+        display: flex;
+        flex-direction: column;
+        gap: var(--gap-uniform);
+
+        span {
+            text-align: center;
+        }
+    }
+}
