+++ Desarrollo web/WebPersonal/src/style-components/shared/icon.css
@@ -1,0 +1,34 @@
+.icon {
+    transition: transform 0.3s linear;
+
+    svg,
+    img,
+    i {
+        width: 100%;
+        height: 100%;
+        display: block;
+    }
+}
+
+.icon:hover {
+    color: #ffffff;
+    transform: scale(1.2);
+}
+
+.icon--social-networking {
+    width: 35px;
+    height: 35px;
+    color: #9ca3af;
+}
+
+.icon-stack {
+    width: 85px;
+    height: 85px;
+}
+
+@media (width < 800px) {
+    .icon-stack {
+        width: 65px;
+        height: 65px;
+    }
+}
