+++ Desarrollo web/WebPersonal/src/style-components/shared/image.css
@@ -1,0 +1,28 @@
+.image {
+    img {
+        display: block;
+        width: 100%;
+        height: 100%;
+    }
+
+    user-select: none;
+}
+
+.image__img--my-photo {
+    width: 380px;
+    height: 380px;
+
+    img {
+        background-color: #a8937e;
+        border-radius: 50%;
+    }
+}
+
+
+
+@media (width < 800px) {
+    .image__img--my-photo {
+        width: 320px;
+        height: 320px;
+    }
+}
